Eliminate race condition that closed server_port too early.
Deferreds yielded by inlineCallbacks are *not* seen by callers of the
decorated function, and inlineCallbacks itself does not maintain
references between yielded Deferreds.  Consequently their lifetimes
are determined by scopes that hold references prior to the point
they're yielded out to inlineCallbacks.  Additionally, the generator's
lifetime is entirely determined by the Deferreds it yields outward and
the variables it closes over.

These two facts together mean that in the following code:

@react
@inlineCallbacks
def f(reactor):
    someObject = object()
    yield Deferred()

...someObject can be deallocated before react even sees the Deferred
returned by inlineCallbacks.

No reference to server_port survived the scope of main(), because it
had been removed from the reactor's reader set.  It was nearly always
deallocated, and its file descriptor closed, after the synchronous
endpoint.listen() call but before HandoffProtocol ran.  In that time
the reactor usually created a pipe for _SIGCHLDWaker that replaced the
now-available listener file descriptor.  This pipe was then sent to
the receiver, which would fail on accept(2) with ENOTSOCK.

Calling dup(2) on the socket created a new file descriptor unrelated
to any Python file-like object, so that it persisted beyond the scope
of main().  An equivalent solution is assign server_port to the
handoff_factory, ensuring that its file descriptor doesn't close when
main() completes.  This also the factory to close the socket on
process shutdown.